Move flashMessagesActions to flashMessageActions. Add Converstaion/Page compoment to route.
