+++ client/src/components/Signup/Page.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
 import Form from './Form'
 import { signupRequest, isAccountExists, isUserExists} from '../../actions/authenticationActions'
-import { addFlashMessage } from '../../actions/flashMessagesActions'
+import { addFlashMessage } from '../../actions/flashMessageActions'
 import FlashMessagesList from '../../flash/FlashMessagesList'
 
 // Localization
